requirements.txt
altgraph==0.17
dis3==0.1.3
et-xmlfile==1.0.1
fbs==0.8.6
future==0.18.2
jdcal==1.4.1
macholib==1.14
numpy==1.18.3
openpyxl==2.6.4
packaging==20.3
pandas==0.25.3
pefile==2019.4.18
Pillow==7.1.1
PyInstaller==3.4
pyparsing==2.4.7
PyQt5==5.9.2
pyqtgraph==0.10.0
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
Shapely==1.7.0
sip==4.19.8
six==1.14.0
xlrd==1.2.0
